requirements.txt
# LangChain
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-postgres>=0.0.15
langchain-text-splitters>=0.3.0

# OpenAI
openai>=1.0.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.0
pgvector>=0.3.0

# PDF loading
pypdf>=4.0.0

# Environment
python-dotenv>=1.0.0
